% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cond_xlsx_utils.R
\name{import_data}
\alias{import_data}
\title{Import data}
\usage{
import_data(path, sheet = NULL)
}
\arguments{
\item{path}{path to file}

\item{sheet}{sheet number or name (if .xlsx)}
}
\value{
imported data
}
\description{
Import data
}
