% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cond_xlsx_utils.R
\name{import_flat_file}
\alias{import_flat_file}
\title{Import flat data file}
\usage{
import_flat_file(path)
}
\arguments{
\item{path}{path to file}
}
\value{
imported flat file
}
\description{
Import flat data file
}
