% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cond_xlsx.R
\name{mod_cond_xlsx_server}
\alias{mod_cond_xlsx_server}
\title{Conditional xlsx panel (Server)}
\usage{
mod_cond_xlsx_server(id)
}
\arguments{
\item{id}{}
}
\value{
shiny server
}
\description{
Conditional xlsx panel (Server)
}
