% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cond_xlsx.R
\name{mod_cond_xlsx_ui}
\alias{mod_cond_xlsx_ui}
\title{Conditional xlsx panel (UI)}
\usage{
mod_cond_xlsx_ui(id)
}
\arguments{
\item{id}{}
}
\value{
shiny UI layout
}
\description{
Conditional xlsx panel (UI)
}
